Query the Matomo Media Analytics for video resources

Returns the media analytics for video resources.

matomo_media_video_titles(
  idSite = NULL,
  period = "day",
  date = "yesterday",
  segment = NULL,
  idSubtable = NULL,
  secondaryDimension = NULL,
  ...
)

Arguments

idSite
The id of your website.

period
The period you request the statistics for. Can be any of: day, week, month, year or range.

date
The date you request the statictics for. Standard format = YYYY-MM-DD or today or yesterday.

segment
Defines the Custom Segment you wish to filter your reports to.

idSubtable
The id of a possible subtable.

secondaryDimension
Secondary dimension is the dimension used in the sub-table of the Event report you are requesting.

...
Possible further options to the query in the format variable = value. For a list of possible options see: https://developer.matomo.org/api-reference/reporting-api

Value

A list containing a data frame with the requested analytics values.

Details

This function expects the hostname and token to be set as environment variables.

Examples

Sys.setenv(MATOMO_HOST = "https://demo.matomo.org")
Sys.setenv(MATOMO_TOKEN = "anonymous")

matomo_media_video_titles(idSite = 62)

#> $content
#>                                                                   label
#> 1                                     Red Sea 2009: Reefs and Sharks HD
#> 2                                                     Indonesia Preview
#> 3                                                    Reefs and Oceanics
#> 4                                                       Similan islands
#> 5                                                       Diving in Coron
#> 6             Diving the Ribbon Reefs on Australia's Great Barrier Reef
#> 7                                             Life around Puerto Galera
#> 8                                                        North Sulawesi
#> 9                                                    Great Barrier Reef
#> 10                                                       Bali Diving HD
#> 11                                    Diving in Tufi (Papua New Guinea)
#> 12                                           Hin Daeng with Lanta Diver
#> 13                                       How to Install Piwik Analytics
#> 14                       Scuba Diving Crocodile Island Boracay Feb 2011
#> 15                         The Russian Destroyer - Shipwrecks of Cayman
#> 16                                       hammerheads at daedalus reef 2
#> 17                                     Komodo - Batu Bolong - Indonésie
#> 18                                                 Muck Diving at Mabul
#> 19                                                        HMAS Brisbane
#> 20                                                             Kona '16
#> 21                    Manta Alley Komodo Indonesia - Amazing HD Footage
#> 22                                                 MV Maldives Princess
#> 23                            Oceans for Youth - Cuba Travel Program-HD
#> 24       Scalloped Hammerhead Shark Highlights 03/05/2009 Layang Layang
#> 25                                           Scuba Diving in Antarctica
#> 26                                             The Murky Side Of Lembeh
#> 27                                          Wakatobi by Full-Moon Light
#> 28                                  West Coast Divers - MV Pawara Promo
#> 29                                                       Dive Galapagos
#> 30                                            Images from Ningaloo Reef
#> 31                                      Piwik Open Source Web Analytics
#> 32                                      Shark Feed | Great Barrier Reef
#> 33                 Cod Feed at Code Hole, Great Barrier Reef, Australia
#> 34                                  Diving in...RAJA AMPAT (West Papua)
#> 35            HD Footage: Diving Malapascua and Gato Island (June 2010)
#> 36          HD Manta Ray footage  - Koh Bon & Similan Islands, Thailand
#> 37      It's Better in Bohol--Diving Balicasag Island, Philippines (HD)
#> 38                                                  Malaysian En-Trance
#> 39                                 North Wreck Safari (Sha'b Abu Nuhas)
#> 40                        Pekko in Drift dive with huge manta in donsol
#> 41                                                     R4 Above & Under
#> 42              Scuba Diving Maldives - Beautiful HD Underwater Footage
#> 43        Sipadan: tropical temptation - Sipadan : tentation tropicale.
#> 44   The uniquely fantastic Infiniti live-aboard in the Andaman Islands
#> 45                                            Blue Hole Arch and Bottom
#> 46                                                     Diving at Anilao
#> 47                                 Diving in and around Nuweiba v1 2011
#> 48                                                  Fiji Aggressor 2016
#> 49                 Form Analytics for Piwik - Improve your online forms
#> 50                Ghosts of the Thistlegorm - Le fantôme du Thistlegorm
#> 51                                       HD 2011 similan & surin diving
#> 52 Infiniti live-aboard's dive trip in the Andaman Islands [March 2016]
#> 53              Introducing Our Newest LiveAboard, the Roatan Aggressor
#> 54                                   Introduction to the Piwik Settings
#> 55                         Mabul Island, Malaysia (1080HD) Travel Video
#> 56                                                              Unknown
#> 57                              MY Blue - Into The Blue - SAILING SOON!
#> 58                     Scuba Diving Malaysia - Pom Pom Island Resort HD
#> 59                                   Sipadan AWESOME Diving - July 2010
#> 60                                        Thresher Sharks at Malapascua
#> 61                                                whale shark encounter
#>    nb_plays nb_unique_visitors_plays nb_impressions
#> 1        27                       16            101
#> 2        21                       21            351
#> 3        13                       12             37
#> 4        12                       12             69
#> 5        11                       10            129
#> 6         9                        7            178
#> 7         9                        9             65
#> 8         8                        8             53
#> 9         7                        7             66
#> 10        6                        6             97
#> 11        6                        6             24
#> 12        5                        5             52
#> 13        5                        5              5
#> 14        5                        5             45
#> 15        5                        5             41
#> 16        4                        4             13
#> 17        4                        4             14
#> 18        4                        4             34
#> 19        3                        3              8
#> 20        3                        3             12
#> 21        3                        3             32
#> 22        3                        1              8
#> 23        3                        3             45
#> 24        3                        3             18
#> 25        3                        3              3
#> 26        3                        3             30
#> 27        3                        3             37
#> 28        3                        2              5
#> 29        2                        2             36
#> 30        2                        2             42
#> 31        2                        2              3
#> 32        2                        2             15
#> 33        1                        1              2
#> 34        1                        1              8
#> 35        1                        1             16
#> 36        1                        1             25
#> 37        1                        1              6
#> 38        1                        1              7
#> 39        1                        1              7
#> 40        1                        1              6
#> 41        1                        1              9
#> 42        1                        1             23
#> 43        1                        1              6
#> 44        1                        1              3
#> 45        0                        0              2
#> 46        0                        0              3
#> 47        0                        0              3
#> 48        0                        0              6
#> 49        0                        0              1
#> 50        0                        0              9
#> 51        0                        0              1
#> 52        0                        0              2
#> 53        0                        0              3
#> 54        0                        0              5
#> 55        0                        0              2
#> 56        0                        0             35
#> 57        0                        0              3
#> 58        0                        0              1
#> 59        0                        0              6
#> 60        0                        0             16
#> 61        0                        0             34
#>    nb_unique_visitors_impressions nb_finishes sum_time_progress
#> 1                              72          14              6603
#> 2                             275           7              3105
#> 3                              34           3              3106
#> 4                              62           2              3683
#> 5                             108           1              2433
#> 6                             151           1              1039
#> 7                              57           4              1070
#> 8                              42           3              1499
#> 9                              60           1               630
#> 10                             73           0              1124
#> 11                             21           3               939
#> 12                             42           1               987
#> 13                              5           1               783
#> 14                             36           1               750
#> 15                             36           2               476
#> 16                             11           3                64
#> 17                             14           2               262
#> 18                             28           0               242
#> 19                              6           0               279
#> 20                             11           1               421
#> 21                             31           1               591
#> 22                              5           3               507
#> 23                             41           2               465
#> 24                             17           1               165
#> 25                              3           1               192
#> 26                             27           2              1124
#> 27                             34           2              1095
#> 28                              4           0               640
#> 29                             30           1                79
#> 30                             37           0               260
#> 31                              3           0               274
#> 32                             15           1               167
#> 33                              2           1                34
#> 34                              8           1               253
#> 35                              9           0               180
#> 36                             23           0               154
#> 37                              6           0                 5
#> 38                              7           0                97
#> 39                              5           0                 5
#> 40                              6           0               171
#> 41                              8           1               161
#> 42                             20           0               164
#> 43                              6           0                43
#> 44                              3           0               159
#> 45                              1           0                 0
#> 46                              3           0                 0
#> 47                              3           0                 0
#> 48                              5           0                 0
#> 49                              1           0                 0
#> 50                              8           0                 0
#> 51                              1           0                 0
#> 52                              2           0                 0
#> 53                              3           0                 0
#> 54                              3           0                 0
#> 55                              2           0                 0
#> 56                             29           0                 0
#> 57                              3           0                 0
#> 58                              1           0                 0
#> 59                              6           0                 0
#> 60                              9           0                 0
#> 61                             33           0                 0
#>    nb_plays_with_tip nb_plays_with_ml sum_fullscreen_plays play_rate
#> 1                 27               27                    0      0.27
#> 2                 21               21                    0      0.06
#> 3                 13               13                    5      0.35
#> 4                 12               12                    0      0.17
#> 5                 11               11                    0      0.09
#> 6                  9                9                    0      0.05
#> 7                  9                9                    0      0.14
#> 8                  8                8                    0      0.15
#> 9                  7                7                    0      0.11
#> 10                 6                6                    0      0.06
#> 11                 6                6                    0      0.25
#> 12                 5                5                    0      0.10
#> 13                 5                5                    3      1.00
#> 14                 5                5                    0      0.11
#> 15                 5                5                    0      0.12
#> 16                 4                4                    0      0.31
#> 17                 4                4                    0      0.29
#> 18                 4                4                    0      0.12
#> 19                 3                3                    0      0.38
#> 20                 3                3                    0      0.25
#> 21                 3                3                    0      0.09
#> 22                 3                3                    0      0.38
#> 23                 3                3                    0      0.07
#> 24                 3                3                    0      0.17
#> 25                 3                3                    0      1.00
#> 26                 3                3                    0      0.10
#> 27                 3                3                    0      0.08
#> 28                 3                3                    0      0.60
#> 29                 2                2                    0      0.06
#> 30                 2                2                    0      0.05
#> 31                 2                2                    2      0.67
#> 32                 2                2                    0      0.13
#> 33                 1                1                    0      0.50
#> 34                 1                1                    1      0.13
#> 35                 1                1                    0      0.06
#> 36                 1                1                    0      0.04
#> 37                 1                1                    0      0.17
#> 38                 1                1                    0      0.14
#> 39                 1                1                    0      0.14
#> 40                 1                1                    0      0.17
#> 41                 1                1                    0      0.11
#> 42                 1                1                    0      0.04
#> 43                 1                1                    0      0.17
#> 44                 1                1                    0      0.33
#> 45                 0                0                    0      0.00
#> 46                 0                0                    0      0.00
#> 47                 0                0                    0      0.00
#> 48                 0                0                    0      0.00
#> 49                 0                0                    0      0.00
#> 50                 0                0                    0      0.00
#> 51                 0                0                    0      0.00
#> 52                 0                0                    0      0.00
#> 53                 0                0                    0      0.00
#> 54                 0                0                    0      0.00
#> 55                 0                0                    0      0.00
#> 56                 0                0                    0      0.00
#> 57                 0                0                    0      0.00
#> 58                 0                0                    0      0.00
#> 59                 0                0                    0      0.00
#> 60                 0                0                    0      0.00
#> 61                 0                0                    0      0.00
#>    finish_rate fullscreen_rate avg_time_watched avg_completion_rate
#> 1         0.52            0.00            18826               55.86
#> 2         0.33            0.00              135                0.44
#> 3         0.23            0.38              176                0.56
#> 4         0.17            0.00              211                0.22
#> 5         0.09            0.00              204                0.44
#> 6         0.11            0.00               58                0.20
#> 7         0.44            0.00              101                0.58
#> 8         0.38            0.00              193                0.64
#> 9         0.14            0.00              102                0.60
#> 10        0.00            0.00              108                0.31
#> 11        0.50            0.00              121                0.72
#> 12        0.20            0.00               69                0.19
#> 13        0.20            0.60              159                0.64
#> 14        0.20            0.00               65                0.23
#> 15        0.40            0.00               74                0.44
#> 16        0.75            0.00               20                0.98
#> 17        0.50            0.00               67                0.82
#> 18        0.00            0.00               62                0.29
#> 19        0.00            0.00               54                0.37
#> 20        0.33            0.00              148                0.84
#> 21        0.33            0.00              154                0.60
#> 22        1.00            0.00          1430918             8466.97
#> 23        0.67            0.00              128                0.81
#> 24        0.33            0.00               58                0.61
#> 25        0.33            0.00               71                0.56
#> 26        0.67            0.00              415                0.84
#> 27        0.67            0.00              369                1.00
#> 28        0.00            0.00              212                0.85
#> 29        0.50            0.00               41                0.79
#> 30        0.00            0.00              136                0.22
#> 31        0.00            1.00              137                0.82
#> 32        0.50            0.00               79                0.74
#> 33        1.00            0.00               35                1.03
#> 34        1.00            1.00              274                1.08
#> 35        0.00            0.00               18                0.08
#> 36        0.00            0.00              124                0.70
#> 37        0.00            0.00                7                0.03
#> 38        0.00            0.00               35                0.20
#> 39        0.00            0.00                7                0.02
#> 40        0.00            0.00               52                0.17
#> 41        1.00            0.00              161                1.00
#> 42        0.00            0.00              165                0.85
#> 43        0.00            0.00               22                0.09
#> 44        0.00            0.00               63                0.16
#> 45        0.00            0.00                0                0.00
#> 46        0.00            0.00                0                0.00
#> 47        0.00            0.00                0                0.00
#> 48        0.00            0.00                0                0.00
#> 49        0.00            0.00                0                0.00
#> 50        0.00            0.00                0                0.00
#> 51        0.00            0.00                0                0.00
#> 52        0.00            0.00                0                0.00
#> 53        0.00            0.00                0                0.00
#> 54        0.00            0.00                0                0.00
#> 55        0.00            0.00                0                0.00
#> 56        0.00            0.00                0                0.00
#> 57        0.00            0.00                0                0.00
#> 58        0.00            0.00                0                0.00
#> 59        0.00            0.00                0                0.00
#> 60        0.00            0.00                0                0.00
#> 61        0.00            0.00                0                0.00
#>    avg_time_to_play avg_media_length
#> 1                13              337
#> 2                98              307
#> 3                40              316
#> 4               113              949
#> 5                41              458
#> 6               224              282
#> 7               171              174
#> 8                19              302
#> 9               170              171
#> 10              119              345
#> 11               69              169
#> 12               22              372
#> 13               11              248
#> 14              307              284
#> 15              443              165
#> 16               50               20
#> 17               15               82
#> 18               38              211
#> 19               48              147
#> 20               11              175
#> 21               50              259
#> 22               84              169
#> 23              269              157
#> 24              206               95
#> 25              169              127
#> 26               19              492
#> 27               26              368
#> 28                9              251
#> 29              231               52
#> 30              207              628
#> 31               14              168
#> 32               22              106
#> 33              137               34
#> 34               61              254
#> 35               49              239
#> 36              128              178
#> 37               14              222
#> 38               38              173
#> 39               25              398
#> 40               30              300
#> 41               46              161
#> 42                9              194
#> 43               31              246
#> 44               11              400
#> 45                0                0
#> 46                0                0
#> 47                0                0
#> 48                0                0
#> 49                0                0
#> 50                0                0
#> 51                0                0
#> 52                0                0
#> 53                0                0
#> 54                0                0
#> 55                0                0
#> 56                0                0
#> 57                0                0
#> 58                0                0
#> 59                0                0
#> 60                0                0
#> 61                0                0
#>                                                                                             segment
#> 1                                                  media_title==Red+Sea+2009%3A+Reefs+and+Sharks+HD
#> 2                                                                    media_title==Indonesia+Preview
#> 3                                                                   media_title==Reefs+and+Oceanics
#> 4                                                                      media_title==Similan+islands
#> 5                                                                      media_title==Diving+in+Coron
#> 6                 media_title==Diving+the+Ribbon+Reefs+on+Australia%26%23039%3Bs+Great+Barrier+Reef
#> 7                                                            media_title==Life+around+Puerto+Galera
#> 8                                                                       media_title==North+Sulawesi
#> 9                                                                   media_title==Great+Barrier+Reef
#> 10                                                                      media_title==Bali+Diving+HD
#> 11                                               media_title==Diving+in+Tufi+%28Papua+New+Guinea%29
#> 12                                                          media_title==Hin+Daeng+with+Lanta+Diver
#> 13                                                      media_title==How+to+Install+Piwik+Analytics
#> 14                                      media_title==Scuba+Diving+Crocodile+Island+Boracay+Feb+2011
#> 15                                        media_title==The+Russian+Destroyer+-+Shipwrecks+of+Cayman
#> 16                                                      media_title==hammerheads+at+daedalus+reef+2
#> 17                                               media_title==Komodo+-+Batu+Bolong+-+Indon%C3%A9sie
#> 18                                                                media_title==Muck+Diving+at+Mabul
#> 19                                                                       media_title==HMAS+Brisbane
#> 20                                                                 media_title==Kona+%26%23039%3B16
#> 21                                   media_title==Manta+Alley+Komodo+Indonesia+-+Amazing+HD+Footage
#> 22                                                                media_title==MV+Maldives+Princess
#> 23                                           media_title==Oceans+for+Youth+-+Cuba+Travel+Program-HD
#> 24                  media_title==Scalloped+Hammerhead+Shark+Highlights+03%2F05%2F2009+Layang+Layang
#> 25                                                          media_title==Scuba+Diving+in+Antarctica
#> 26                                                            media_title==The+Murky+Side+Of+Lembeh
#> 27                                                         media_title==Wakatobi+by+Full-Moon+Light
#> 28                                                 media_title==West+Coast+Divers+-+MV+Pawara+Promo
#> 29                                                                      media_title==Dive+Galapagos
#> 30                                                           media_title==Images+from+Ningaloo+Reef
#> 31                                                     media_title==Piwik+Open+Source+Web+Analytics
#> 32                                                   media_title==Shark+Feed+%7C+Great+Barrier+Reef
#> 33                            media_title==Cod+Feed+at+Code+Hole%2C+Great+Barrier+Reef%2C+Australia
#> 34                                             media_title==Diving+in...RAJA+AMPAT+%28West+Papua%29
#> 35                     media_title==HD+Footage%3A+Diving+Malapascua+and+Gato+Island+%28June+2010%29
#> 36               media_title==HD+Manta+Ray+footage++-+Koh+Bon+%26amp%3B+Similan+Islands%2C+Thailand
#> 37    media_title==It%26%23039%3Bs+Better+in+Bohol--Diving+Balicasag+Island%2C+Philippines+%28HD%29
#> 38                                                                 media_title==Malaysian+En-Trance
#> 39                                 media_title==North+Wreck+Safari+%28Sha%26%23039%3Bb+Abu+Nuhas%29
#> 40                                       media_title==Pekko+in+Drift+dive+with+huge+manta+in+donsol
#> 41                                                            media_title==R4+Above+%26amp%3B+Under
#> 42                             media_title==Scuba+Diving+Maldives+-+Beautiful+HD+Underwater+Footage
#> 43                   media_title==Sipadan%3A+tropical+temptation+-+Sipadan+%3A+tentation+tropicale.
#> 44                  media_title==The+uniquely+fantastic+Infiniti+live-aboard+in+the+Andaman+Islands
#> 45                                                           media_title==Blue+Hole+Arch+and+Bottom
#> 46                                                                    media_title==Diving+at+Anilao
#> 47                                                media_title==Diving+in+and+around+Nuweiba+v1+2011
#> 48                                                                 media_title==Fiji+Aggressor+2016
#> 49                                media_title==Form+Analytics+for+Piwik+-+Improve+your+online+forms
#> 50                          media_title==Ghosts+of+the+Thistlegorm+-+Le+fant%C3%B4me+du+Thistlegorm
#> 51                                              media_title==HD+2011+similan+%26amp%3B+surin+diving
#> 52 media_title==Infiniti+live-aboard%26%23039%3Bs+dive+trip+in+the+Andaman+Islands+%5BMarch+2016%5D
#> 53                           media_title==Introducing+Our+Newest+LiveAboard%2C+the+Roatan+Aggressor
#> 54                                                  media_title==Introduction+to+the+Piwik+Settings
#> 55                                  media_title==Mabul+Island%2C+Malaysia+%281080HD%29+Travel+Video
#> 56                                                                             media_title==Unknown
#> 57                                           media_title==MY+Blue+-+Into+The+Blue+-+SAILING+SOON%21
#> 58                                    media_title==Scuba+Diving+Malaysia+-+Pom+Pom+Island+Resort+HD
#> 59                                                  media_title==Sipadan+AWESOME+Diving+-+July+2010
#> 60                                                       media_title==Thresher+Sharks+at+Malapascua
#> 61                                                               media_title==whale+shark+encounter
#>    idsubdatatable
#> 1             241
#> 2             259
#> 3             175
#> 4             229
#> 5             247
#> 6             253
#> 7             217
#> 8             211
#> 9             223
#> 10            235
#> 11            133
#> 12            205
#> 13             25
#> 14            199
#> 15            181
#> 16             97
#> 17            103
#> 18            157
#> 19             67
#> 20             91
#> 21            151
#> 22             73
#> 23            193
#> 24            121
#> 25             19
#> 26            145
#> 27            169
#> 28             31
#> 29            163
#> 30            187
#> 31             13
#> 32            109
#> 33              1
#> 34             79
#> 35            115
#> 36            139
#> 37             49
#> 38             55
#> 39             61
#> 40             43
#> 41             85
#> 42            127
#> 43             37
#> 44              7
#> 45             NA
#> 46             NA
#> 47             NA
#> 48             NA
#> 49             NA
#> 50             NA
#> 51             NA
#> 52             NA
#> 53             NA
#> 54             NA
#> 55             NA
#> 56             NA
#> 57             NA
#> 58             NA
#> 59             NA
#> 60             NA
#> 61             NA
#> 
#> $url
#> [1] "https://demo.matomo.org/index.php?module=API&method=MediaAnalytics.getVideoTitles&idSite=62&period=day&date=yesterday"
#> 
#> $response
#> Response [https://demo.matomo.org/index.php?module=API&method=MediaAnalytics.getVideoTitles&idSite=62&period=day&date=yesterday&format=JSON&token_auth=anonymous]
#>   Date: 2019-11-13 14:59
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 29 kB
#> 
#> 
#> attr(,"class")
#> [1] "matomo_api"